+++ pennylane_lightning/core/simulators/lightning_qubit/catalyst/LightningSimulator.cpp
@@ -177,7 +177,8 @@ void LightningSimulator::NamedOperation(
     const std::string &name, const std::vector<double> &params,
     const std::vector<QubitIdType> &wires, bool inverse,
     const std::vector<QubitIdType> &controlled_wires,
-    const std::vector<bool> &controlled_values) {
+    const std::vector<bool> &controlled_values,
+    [[maybe_unused]] const std::vector<std::string> &optional_params) {
     // Check the validity of number of qubits and parameters
     RT_FAIL_IF(controlled_wires.size() != controlled_values.size(),
                "Controlled wires/values size mismatch");
@@ -188,6 +189,20 @@ void LightningSimulator::NamedOperation(
     // Convert wires to device wires
     auto &&dev_wires = getDeviceWires(wires);
     auto &&dev_controlled_wires = getDeviceWires(controlled_wires);
+
+    if (name == "PauliRot") {
+        RT_FAIL_IF(optional_params.size() != 1,
+                   "PauliRot operation requires one string "
+                   "parameter for the Pauli word");
+        RT_FAIL_IF(!controlled_wires.empty(),
+                   "Controlled PauliRot is not supported");
+        RT_FAIL_IF(this->tape_recording,
+                   "PauliRot operation is not supported when tape "
+                   "recording is active"); // TODO: support caching
+        this->device_sv->applyPauliRot(dev_wires, inverse, params,
+                                       optional_params[0]);
+        return;
+    }
 
     // Update the state-vector
     if (controlled_wires.empty()) {