+++ .github/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 <h3>New features since last release</h3>
 
+- Added support for `qml.PauliRot` simulation for Catalyst.
+  [(#1322)](https://github.com/PennyLaneAI/pennylane-lightning/pull/1322)
+
 <h3>Improvements 🛠</h3>
 
 - Added a C++ kernel for the `qml.PauliRot` simulation on `lightning.gpu`.
@@ -40,6 +43,7 @@
 This release contains contributions from (in alphabetical order):
 
 Ali Asadi,
+Jeffrey Kam,
 Jake Zaia.
 
 ---